finrobot/requirements.txt
python-dateutil==2.9.0.post0
tqdm==4.66.5
joblib==1.4.2
scikit-learn==1.5.2

transformers==4.44.2
sentence-transformers==3.0.1

feedparser==6.0.11
praw==7.7.1
yfinance==0.2.43

finnhub-python==2.4.20
fmp-python

matplotlib==3.9.2
seaborn==0.13.2
mplfinance==0.12.10b0

requests==2.32.3
huggingface_hub==0.24.0
